initial pose

+++ src/main/java/ch/ethz/idsc/retina/demo/jph/LogEventExtract.java
@@ -4,7 +4,8 @@ package ch.ethz.idsc.retina.demo.jph;
 import java.io.File;
 
 import ch.ethz.idsc.retina.demo.DubendorfHangarLog;
-import ch.ethz.idsc.subare.util.UserHome;
+import ch.ethz.idsc.retina.lcm.OfflineLogPlayer;
+import idsc.BinaryBlob;
 import lcm.logging.Log;
 import lcm.logging.Log.Event;
 import lcm.logging.LogEventWriter;
@@ -16,12 +17,13 @@ enum LogEventExtract {
     File src = new File("/media/datahaki/media/ethz/gokartlogs", "20180112T113153_9e1d3699.lcm.00");
     src = DubendorfHangarLog._20180112T154355_9e1d3699.file(GokartLcmLogPlayer.LOG_ROOT);
     // new File("/media/datahaki/mobile/temp", "20180108T162528_5f742add.lcm.00");
-    File dst = UserHome.file("20180108T165210_maxtorque.lcm");
-    // new File("/home/datahaki/Projects/retina/src/test/resources/localization", "Xvlp16.center.pos.lcm");
-    dst = UserHome.file("20180112T154355.lcm");
-    dst = new File("/home/datahaki/gokart/pursuit/20180112T154355/log.lcm");
-    dst.delete();
-    int lo = 0;
+    File dst;
+    dst = new File("/home/datahaki/attemptlog.lcm");
+    if (dst.exists()) {
+      System.out.println("deleting: " + dst);
+      dst.delete();
+    }
+    int lo = 3089651;
     int hi = 3816812;
     // ---
     Log log = new Log(src.toString(), "r");
@@ -30,13 +32,23 @@ enum LogEventExtract {
       // int count = 0;
       while (true) {
         Event event = log.readNext();
-        if (lo <= event.eventNumber && event.eventNumber < hi)
-          logWriter.write(event);
+        if (lo <= event.eventNumber && event.eventNumber < hi) {
+          try {
+            new BinaryBlob(event.data);
+            logWriter.write(event);
+          } catch (Exception exception) {
+            // ---
+            exception.printStackTrace();
+          }
+        }
       }
     } catch (Exception exception) {
       System.err.println(exception.getMessage());
       // ---
     }
     logWriter.close();
+    // ---
+    System.out.println("check consistency");
+    OfflineLogPlayer.process(dst, MessageConsistency.INSTANCE);
   }
 }
